refactor(engine): extract trading-cycle orchestration
Lift the per-cycle orchestration cluster out of TradingEngine
(strategy.py) into engines/cycle_runner.py, mirroring the #1062
WS-health pattern: module-level functions that take `engine`, with a
thin `_cycle` delegate left on the engine.

Pure mechanical move — no behavior change. The runner reads/writes only
existing engine attributes/methods; calls to engine-resident primitives
(_fetch_positions, _audit_orders, _fetch_total_equity,
_positions_to_status_format, _broker_calls, _process_symbol) stay routed
through `engine.`. The order-submission seam (_process_symbol,
_handle_decision, _validate_and_place_order) is untouched.

- New engines/cycle_runner.py: run_cycle(engine) + cluster helpers
  (_run_cycle_inner, _report_system_status, _fetch_positions_and_audit,
  _await_audit_task, _compute_equity, _update_equity_and_risk,
  _fetch_batch_tickers).
- strategy.py: 2810 -> 2635 lines; _cycle reduced to a one-line
  delegate; drops now-unused correlation_context / trace_span imports.

Continues the seam-by-seam decomposition (after #1062).
